+++ isy-sonderzeichen/src/main/java/de/bund/bva/pliscommon/plissonderzeichen/dinspec91379/konstanten/TransformationsKonstanten.java
@@ -18,7 +18,6 @@ package de.bund.bva.pliscommon.plissonderzeichen.dinspec91379.konstanten;
 
 /**
  * This class contains constants for the resources of the respective transformation table.
- *
  */
 public class TransformationsKonstanten {
 
@@ -40,4 +39,10 @@ public class TransformationsKonstanten {
 
     /** The path to the transformation transcription table. */
     public static final String TRANSFORMATIONS_TABELLE_TRANSKRIPTION = "/resources/tabellen/transformation_din91379_transkription.transform";
+
+    /** The path to the legacy transcription table, tranforming DIN SPEC 91379 characters to StringLatin1.1. */
+    public static final String TRANSFORMATIONS_TABELLE_LEGACY = "/resources/tabellen/transformation_dinspec91379_zu_1_1.transform";
+
+    private TransformationsKonstanten() {
+    }
 }