+++ isy-sonderzeichen/CHANGELOG.md
@@ -2,6 +2,9 @@
 - `IFS-1003`:
   + Refactoring des DIN-SPEC-Packages von isy-sonderzeichen
   + Entfernen des `core`-Unterpakets des DIN-SPEC-Packages von isy-sonderzeichen
+- `IFS-1109`: DIN SPEC 91379:
+  * Erweiterung Transformator-Interface um Methoden, die Metadaten der Transformation tracken
+  * neuer LegacyTransformator, der Texte nach DIN-SPEC-91379 in String.Latin-1.1-konforme Texte transformiert
   
 # 1.10.0:
 - `IFS-849`: